+++ tests/Games/NexusMods.Games.TestFramework/Downloader/FileSystemModMetadata.cs
@@ -1,7 +1,7 @@
 ﻿using System.Text.Json.Serialization;
 using NexusMods.Paths;
 
-namespace NexusMods.Games.TestFramework.Stubs;
+namespace NexusMods.Games.TestFramework.Downloader;
 
 /// <summary>
 /// Metadata used for downloading mods from the FileSystem.
