+++ tests/Games/NexusMods.Games.TestFramework/Downloader/RemoteModSource.cs
@@ -1,4 +1,4 @@
-﻿namespace NexusMods.Games.TestFramework.Stubs;
+﻿namespace NexusMods.Games.TestFramework.Downloader;
 
 /// <summary>
 /// Defines the source this test mod is from.
